+++ R/site.R
@@ -1,7 +1,7 @@
 blogdown_site = function(input, ...) {
   # start serving the site when a blogdown project is opened in RStudio
   if (interactive() && getOption('blogdown.serve_site.startup', TRUE))
-    try(preview_site())
+    try(preview_site(startup = TRUE))
 
   output_dir = publish_dir()
   render = function(input_file, output_format, envir, quiet, encoding, ...) {